+++ composer.json
@@ -27,13 +27,25 @@
 		"ext-gd" : "*",
 		"ext-mbstring" : "*",
 		"ext-xml" : "*",
-		"sabre/dav" : "~3.2",
-		"ext-openssl" : "*"
+		"ext-openssl" : "*",
+		"sabre/dav" : "~3.2"
+	},
+	"require-dev" : {
+		"php" : ">=5.6",
+		"phpunit/phpunit" : "^5.6",
+		"behat/behat" : "@stable",
+		"behat/mink-extension": "@stable",
+		"behat/mink-goutte-driver": "@stable"
 	},
 	"autoload" : {
 		"psr-4" : {
 			"Hubzilla\\" : "include/",
 			"Zotlabs\\" : "Zotlabs/"
+		}
+	},
+	"autoload-dev" : {
+		"psr-4" : {
+			"Zotlabs\\Tests\\Unit\\" : "tests/unit"
 		}
 	},
 	"minimum-stability" : "stable",
