+++ tests/unit/TextTest.php
@@ -1,20 +1,22 @@
 <?php
 /**
  * this file contains tests for text.php
- * 
+ *
  * @package test.util
  */
+
+use PHPUnit\Framework\TestCase;
 
 /** required, it is the file under test */
 require_once('include/text.php');
 
 /**
  * TestCase for the texter
- * 
+ *
  * @author ken restivo
  * @package test.util
  */
-class TextTest extends PHPUnit_Framework_TestCase {
+class TextTest extends TestCase {
 	public function testGoodEmail() {
 		$this->assertTrue(valid_email_regex('ken@spaz.org'));
 	}
