+++ tests/unit/template_test.php
